+++ src/components/DatePicker/DatePickerTypeMonthRange/DatePickerTypeMonthRange.tsx
@@ -9,7 +9,6 @@ import React, {
 } from 'react';
 
 import { useClickOutside } from '../../../hooks/useClickOutside/useClickOutside';
-import { useFlag } from '../../../hooks/useFlag/useFlag';
 import { useForkRef } from '../../../hooks/useForkRef/useForkRef';
 import {
   DatePickerDropdown,
@@ -21,11 +20,13 @@ import {
   datePickerPropDateTimeViewDefault,
   DatePickerTypeComponent,
 } from '../types';
+import { useCalendarVisible } from '../useCalendarVisible';
 import { useCurrentVisibleDate } from '../useCurrentVisibleDate';
 
 export const DatePickerTypeMonthRange: DatePickerTypeComponent<'month-range'> =
   forwardRef((props, ref) => {
     const {
+      onChange,
       events,
       dateTimeView = datePickerPropDateTimeViewDefault,
       locale,
@@ -40,6 +41,9 @@ export const DatePickerTypeMonthRange: DatePickerTypeComponent<'month-range'> =
       renderAdditionalControls,
       inputRef,
       name,
+      onDropdownOpen,
+      dropdownOpen,
+      ignoreOutsideClicksRefs,
       ...fieldProps
     } = props;
 
@@ -71,7 +75,12 @@ export const DatePickerTypeMonthRange: DatePickerTypeComponent<'month-range'> =
       }
     };
 
-    const [calendarVisible, setCalendarVisible] = useFlag(false);
+    const [calendarVisible, setCalendarVisible] = useCalendarVisible({
+      dropdownOpen,
+      onDropdownOpen,
+      startRef: startFieldInputRef,
+      endRef: endFieldInputRef,
+    });
 
     const [currentVisibleDate, setCurrentVisibleDate] = useCurrentVisibleDate({
       currentVisibleDate: currentVisibleDateProp,
@@ -91,13 +100,11 @@ export const DatePickerTypeMonthRange: DatePickerTypeComponent<'month-range'> =
 
     const startFieldOnFocusHandler = (e: React.FocusEvent<HTMLElement>) => {
       setFieldFocused('start');
-      setCalendarVisible.on();
       Array.isArray(onFocus) ? onFocus[0]?.(e) : onFocus?.(e);
     };
 
     const endFieldOnFocusHandler = (e: React.FocusEvent<HTMLElement>) => {
       setFieldFocused('end');
-      setCalendarVisible.on();
       Array.isArray(onFocus) ? onFocus[1]?.(e) : onFocus?.(e);
     };
 
@@ -144,7 +151,12 @@ export const DatePickerTypeMonthRange: DatePickerTypeComponent<'month-range'> =
 
     useClickOutside({
       isActive: calendarVisible,
-      ignoreClicksInsideRefs: [startFieldRef, endFieldRef, calendarRef],
+      ignoreClicksInsideRefs: [
+        startFieldRef,
+        endFieldRef,
+        calendarRef,
+        ...(ignoreOutsideClicksRefs ?? []),
+      ],
       handler: useCallback(() => {
         setFieldFocused(false);
         setCalendarVisible.off();
@@ -158,6 +170,10 @@ export const DatePickerTypeMonthRange: DatePickerTypeComponent<'month-range'> =
           ref={ref}
           startFieldRef={startFieldRef}
           endFieldRef={endFieldRef}
+          onChange={(date) => {
+            onChange?.(date);
+            setCalendarVisible.on();
+          }}
           startFieldInputRef={useForkRef([startFieldInputRef, inputRef?.[0]])}
           endFieldInputRef={useForkRef([endFieldInputRef, inputRef?.[1]])}
           startFieldOnFocus={startFieldOnFocusHandler}
@@ -176,6 +192,8 @@ export const DatePickerTypeMonthRange: DatePickerTypeComponent<'month-range'> =
           endFieldOnBlur={endFieldOnBlurHandler}
           startFocused={startFocused}
           endFocused={endFocused}
+          startFieldOnClick={setCalendarVisible.on}
+          endFieldOnClick={setCalendarVisible.on}
           startFieldName={Array.isArray(name) ? name[0] : `${name}_start`}
           endFieldName={Array.isArray(name) ? name[1] : `${name}_end`}
         />