+++ src/components/DatePicker/DatePickerTypeTime/DatePickerTypeTime.tsx
@@ -1,7 +1,6 @@
 import React, { forwardRef, useEffect, useRef } from 'react';
 
 import { useClickOutside } from '../../../hooks/useClickOutside/useClickOutside';
-import { useFlag } from '../../../hooks/useFlag/useFlag';
 import { setRef } from '../../../utils/setRef';
 import { DatePickerDropdown } from '../DatePickerDropdown/DatePickerDropdown';
 import { DatePickerFieldTypeTime } from '../DatePickerFieldTypeTime/DatePickerFieldTypeTime';
@@ -14,23 +13,26 @@ import {
   datePickerPropDateTimeViewDefault,
   DatePickerTypeComponent,
 } from '../types';
+import { useCalendarVisible } from '../useCalendarVisible';
 
 export const DatePickerTypeTime: DatePickerTypeComponent<'time'> = forwardRef(
   (props, ref) => {
     const {
+      onChange,
       events,
       dateTimeView = datePickerPropDateTimeViewDefault,
       locale,
       dropdownForm,
       dropdownClassName,
-      onFocus,
       multiplicityHours: multiplicityHoursProp,
       multiplicityMinutes: multiplicityMinutesProp,
       multiplicitySeconds: multiplicitySecondsProp,
       renderAdditionalControls,
-
       currentVisibleDate,
       onChangeCurrentVisibleDate,
+      onDropdownOpen,
+      dropdownOpen,
+      ignoreOutsideClicksRefs,
       ...otherProps
     } = props;
 
@@ -44,13 +46,13 @@ export const DatePickerTypeTime: DatePickerTypeComponent<'time'> = forwardRef(
 
     const fieldRef = useRef<HTMLDivElement>(null);
     const calendarRef = useRef<HTMLDivElement>(null);
+    const inputRef = useRef<HTMLInputElement>(null);
 
-    const [calendarVisible, setCalendarVisible] = useFlag(false);
-
-    const onFocusHandler = (e: React.FocusEvent<HTMLElement>) => {
-      onFocus && onFocus(e);
-      setCalendarVisible.on();
-    };
+    const [calendarVisible, setCalendarVisible] = useCalendarVisible({
+      dropdownOpen,
+      onDropdownOpen,
+      startRef: inputRef,
+    });
 
     useEffect(() => {
       if (ref) {
@@ -60,7 +62,11 @@ export const DatePickerTypeTime: DatePickerTypeComponent<'time'> = forwardRef(
 
     useClickOutside({
       isActive: calendarVisible,
-      ignoreClicksInsideRefs: [fieldRef, calendarRef],
+      ignoreClicksInsideRefs: [
+        fieldRef,
+        calendarRef,
+        ...(ignoreOutsideClicksRefs ?? []),
+      ],
       handler: setCalendarVisible.off,
     });
 
@@ -69,7 +75,12 @@ export const DatePickerTypeTime: DatePickerTypeComponent<'time'> = forwardRef(
         <DatePickerFieldTypeTime
           {...otherProps}
           ref={fieldRef}
-          onFocus={onFocusHandler}
+          inputRef={inputRef}
+          onChange={(date) => {
+            onChange?.(date);
+            setCalendarVisible.on();
+          }}
+          onClick={setCalendarVisible.on}
           multiplicityHours={multiplicityHours}
           multiplicitySeconds={multiplicitySeconds}
           multiplicityMinutes={multiplicityMinutes}